Cleanup

+++ TewlKit/Asm/Stubs/IndirectJumpEaxStub.cs
@@ -19,7 +19,7 @@ namespace TewlKit.Asm.Stubs
             // Opcode for MOV EAX, imm32 is 0xB8
             stub.Write(OpCodes.MovEax);
             // Write the 32-bit address to jump to
-            stub.Write((uint)address.ToInt32());
+            stub.Write(address.ToInt32());
             // Opcode for JMP EAX is 0xFFE0
             stub.Write(OpCodes.JmpEaxRax);
             // Return the generated stub as a byte array
